fix an implicit cast to float

+++ interface/src/Application.cpp
@@ -3108,7 +3108,7 @@ void Application::updateWindowTitle(){
     
     AccountManager& accountManager = AccountManager::getInstance();
     if (accountManager.getAccountInfo().hasBalance()) {
-        float creditBalance = accountManager.getAccountInfo().getBalance() / SATOSHIS_PER_CREDIT;
+        float creditBalance = accountManager.getAccountInfo().getBalance() / (float) SATOSHIS_PER_CREDIT;
         
         QString creditBalanceString;
         creditBalanceString.sprintf("%.8f", creditBalance);
